feat(frontend): add cinematic landing page
